+++ internal/home/control.go
@@ -103,9 +103,8 @@ type statusResponse struct {
 	DNSPort             int      `json:"dns_port"`
 	HTTPPort            int      `json:"http_port"`
 	IsProtectionEnabled bool     `json:"protection_enabled"`
-	// ProtectionDisabledUntil is a timestamp until when the protection is
-	// disabled.
-	ProtectionDisabledUntil *time.Time `json:"protection_disabled_until"`
+	// ProtectionDisabledDuration is a pause duration in milliseconds.
+	ProtectionDisabledDuration int64 `json:"protection_disabled_duration"`
 	// TODO(e.burkov): Inspect if front-end doesn't requires this field as
 	// openapi.yaml declares.
 	IsDHCPAvailable bool `json:"dhcp_available"`
@@ -135,21 +134,20 @@ func handleStatus(w http.ResponseWriter, r *http.Request) {
 		config.RLock()
 		defer config.RUnlock()
 
-		var disabledUntil *time.Time
-		if config.DNS.ProtectionDisabledUntil != nil {
-			t := *config.DNS.ProtectionDisabledUntil
-			disabledUntil = &t
+		var pauseDuration int64
+		if until := config.DNS.ProtectionDisabledUntil; until != nil {
+			pauseDuration = time.Until(*until).Milliseconds()
 		}
 
 		resp = statusResponse{
-			Version:                 version.Version(),
-			DNSAddrs:                dnsAddrs,
-			DNSPort:                 config.DNS.Port,
-			HTTPPort:                config.BindPort,
-			Language:                config.Language,
-			IsRunning:               isRunning(),
-			ProtectionDisabledUntil: disabledUntil,
-			IsProtectionEnabled:     isProtectionEnabled,
+			Version:                    version.Version(),
+			DNSAddrs:                   dnsAddrs,
+			DNSPort:                    config.DNS.Port,
+			HTTPPort:                   config.BindPort,
+			Language:                   config.Language,
+			IsRunning:                  isRunning(),
+			ProtectionDisabledDuration: pauseDuration,
+			IsProtectionEnabled:        isProtectionEnabled,
 		}
 	}()