+++ openapi/CHANGELOG.md
@@ -13,9 +13,9 @@
 * The new field `"protection_disabled_until"` in `GET /control/dns_info` is the
   timestamp until when the protection is disabled.
 
-### New `"protection_disabled_until"` field in `GET /control/status` response
+### New `"protection_disabled_duration"` field in `GET /control/status` response
 
-* The new field `"protection_disabled_until"` is the duration of protection
+* The new field `"protection_disabled_duration"` is the duration of protection
   pause in milliseconds.
 
 ### `POST /control/protection`